+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 =========
 
 * Some tweaks to `.JSONActor` and the testing framework.
+* Added an error reply level.
 
 * :release:`0.1.11 <2020-01-14>`
 * Remove numpy dependency from CLU.